+++ src/main/java/kmlib/console/ListFactionsCommand.java
@@ -11,9 +11,9 @@ import kmlib.starsector.colonies.Colony;
 import kmlib.starsector.colonies.SectorColonies;
 import kmlib.starsector.factions.FactionFlags;
 import kmlib.starsector.factions.StarsectorPlayerFactionResolver;
+import kmlib.starsector.factions.relation.StarsectorPlayerRelations;
+import kmlib.starsector.factions.relation.StarsectorRelationFormatter;
 import kmlib.starsector.markets.MarketVisibility;
-import kmlib.starsector.relation.StarsectorPlayerRelations;
-import kmlib.starsector.relation.StarsectorRelationFormatter;
 
 import java.util.ArrayList;
 import java.util.Comparator;